hyuna cleaned up

+++ cogs/! GGS.py
@@ -407,14 +407,9 @@ class GGS(commands.Cog):
 
     @commands.command()
     async def hyuna(self, ctx):
-        if ctx.guild.id == luminary:
-            if ctx.channel.id == kbotcom:
-                await ctx.send(f'<@!{ctx.author.id}> is talking about Hyuna :heart: ')
-                await ctx.send(random.choice(self.bot.hyuna_gif))
-                await ctx.message.delete()
-            else:
-                await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
-                await ctx.message.delete()
+        if ctx.guild.id == luminary and ctx.channel.id != kbotcom:
+            await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
+            await ctx.message.delete()
         else:
             await ctx.send(f'<@!{ctx.author.id}> is talking about Hyuna :heart:')
             await ctx.send(random.choice(self.bot.hyuna_gif))
